Update RedisPubSubSpout.java
Correct me if I'm wrong but each instance of the Spout is creating their own Redis pool and each Thread is getting just one resource of this pool. Said that, why don't use a regular Jedis instance in place of the JedisPool one?

I was looking into the code and I don't view any reason to have the JedisPool and if you just use the RedisInstance (that usually will go in a Singleton Factory) you can save a bunch of resources.

Let me know your thoughts and thanks for the good work. This class is really useful.